backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
redis==5.0.8
pydantic==2.9.2
pydantic-settings==2.5.2
